binance/requirements.txt
#everything that we need
python-binance
TA-Lib
numpy
websocket_client